added yup validations
