+++ app/javascript/src/components/constants.js
@@ -11,44 +11,3 @@ export const FEATURE_LIST = [
   "Performs background job processing 'inline' for heroku env. It means heroku can deliver emails.",
   "Letter opener gem for development.",
 ];
-
-export const DUMMY_CONTACTS = [
-  {
-    label: "Oliver Smith",
-    value: "Oliver Smith",
-    id: 1,
-  },
-  {
-    label: "Tony Stark",
-    value: "Tony Stark",
-    id: 2,
-  },
-  {
-    label: "Sherlock Holmes",
-    value: "Tony Stark",
-    id: 3,
-  },
-  {
-    label: "William Butcher",
-    value: "Tony Stark",
-    id: 4,
-  },
-];
-
-export const DUMMY_TAGS = [
-  {
-    label: "Sales",
-    value: "sales",
-    id: 1,
-  },
-  {
-    label: "Finance",
-    value: "finance",
-    id: 2,
-  },
-  {
-    label: "User Experience",
-    value: "userExperience",
-    id: 3,
-  },
-];
